+++ database/migrations/2014_12_03_230347_create_lt_Category_table.php
@@ -61,7 +61,7 @@ class CreateLtCategoryTable extends Migration
                 ->onUpdate('cascade')
                 ->onDelete('cascade');
 
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->foreign('user_id')
                 ->references('id')
                 ->on('users')
